Refactor data structure; Implement TabsDashboard for better navigation; Remove unused components and files; Consolidate utility functions.
