+++ app/dashboard/tabs/views/top-chart.tsx
