+++ app/dashboard/tabs/views/tops.tsx
@@ -7,12 +7,7 @@ import {
   TooltipTrigger,
 } from "@radix-ui/react-tooltip";
 import { useEffect, useRef, useState } from "react";
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-} from "../../../components/ui/card";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { cn } from "@/lib/utils";
 
 export default function Tops({
