+++ auth.ts
@@ -2,7 +2,7 @@ import { compare } from "bcrypt-ts";
 import NextAuth from "next-auth";
 import Credentials from "next-auth/providers/credentials";
 import { z } from "zod";
-import { getUser } from "./app/lib/db";
+import { getUser } from "./lib/db";
 import { authConfig } from "./auth.config";
 
 export const { auth, signIn, signOut } = NextAuth({
